refactor: consolidate user event schemas into shared models package
