+++ packages/models/src/user/user.ts
@@ -23,3 +23,67 @@ export const User = z.object({
   productRole: UserRole,
 });
 export type User = z.infer<typeof User>;
+
+export const selfcareUserEventType = {
+  add: "ADD",
+  update: "UPDATE",
+} as const;
+
+export const SelfcareUserEventType = z.enum([
+  selfcareUserEventType.add,
+  selfcareUserEventType.update,
+]);
+export type SelfcareUserEventType = z.infer<typeof SelfcareUserEventType>;
+
+export const relationshipStatus = {
+  active: "ACTIVE",
+  suspended: "SUSPENDED",
+  deleted: "DELETED",
+  rejected: "REJECTED",
+} as const;
+export const RelationshipStatus = z.enum([
+  Object.values(relationshipStatus)[0],
+  ...Object.values(relationshipStatus).slice(1),
+]);
+export type RelationshipStatus = z.infer<typeof RelationshipStatus>;
+
+const BaseUser = z.object({
+  userId: z.string().uuid().nullish(),
+  name: z.string(),
+  familyName: z.string(),
+  email: z.string(),
+  role: z.string(),
+  productRole: UserRole,
+  mobilePhone: z.string().nullish(),
+});
+
+// Base event with common fields
+const BaseEvent = z.object({
+  id: z.string(),
+  institutionId: z.string().trim().min(1), // Selfcare ID
+  productId: z.string().trim().min(1),
+});
+
+// Add user event (only with active status)
+const AddUserEvent = BaseEvent.extend({
+  eventType: z.literal(selfcareUserEventType.add),
+  user: BaseUser.extend({
+    relationshipStatus: z.literal(relationshipStatus.active),
+  }),
+});
+
+// Update user event (can be active or deleted status)
+const UpdateUserEvent = BaseEvent.extend({
+  eventType: z.literal(selfcareUserEventType.update),
+  user: BaseUser.extend({
+    relationshipStatus: z.union([
+      z.literal(relationshipStatus.active),
+      z.literal(relationshipStatus.deleted),
+    ]),
+  }),
+});
+
+// Combined schema
+// Base payload type before transformation
+export const BaseUsersEventPayload = z.union([AddUserEvent, UpdateUserEvent]);
+export type BaseUsersEventPayload = z.infer<typeof BaseUsersEventPayload>;
